Convert simple API collections to OO controllers
Collections that neither are a subcollection nor have subcollections can
be easily converted to OO-type controllers. The remaining collections
that do not fall into this category are more problematic, so will be
done in future revisions.

The converted collections are:

* arbitration profiles
* arbitration rules
* arbitration profiles
* arbitration rules
* arbitration settings
* availability zones
* container deployments
* currencies
* flavors
* instances
* measures
* pictures
* provision dialogs
* requests
* security groups
* servers
* virtual templates
* zones
