+++ app/controllers/api/arbitration_settings_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class ArbitrationSettingsController < BaseController
+  end
+end
