+++ app/controllers/api/availability_zones_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class AvailabilityZonesController < BaseController
+  end
+end
