+++ app/controllers/api/base_controller.rb
@@ -60,28 +60,23 @@ module Api
     include_concern 'Generic'
 
     include_concern 'Accounts'
-    include_concern 'ArbitrationProfiles'
-    include_concern 'ArbitrationRules'
     include_concern 'Authentication'
     include_concern 'AutomationRequests'
     include_concern 'Blueprints'
     include_concern 'Categories'
     include_concern 'CloudNetworks'
     include_concern 'Conditions'
-    include_concern 'ContainerDeployments'
     include_concern 'CustomAttributes'
     include_concern 'Events'
     include_concern 'Features'
     include_concern 'Groups'
     include_concern 'Hosts'
-    include_concern 'Instances'
     include_concern 'Policies'
     include_concern 'PolicyActions'
     include_concern 'Providers'
     include_concern 'ProvisionRequests'
     include_concern "Rates"
     include_concern "Reports"
-    include_concern 'Requests'
     include_concern 'RequestTasks'
     include_concern 'ResourceActions'
     include_concern 'Roles'
