--- app/controllers/api/base_controller/requests.rb
@@ -1,16 +1,0 @@
-module Api
-  class BaseController
-    module Requests
-      def find_requests(id)
-        klass = collection_class(:requests)
-        return klass.find(id) if @auth_user_obj.admin?
-        klass.find_by!(:requester => @auth_user_obj, :id => id)
-      end
-
-      def requests_search_conditions
-        return {} if @auth_user_obj.admin?
-        {:requester => @auth_user_obj}
-      end
-    end
-  end
-end
