+++ app/controllers/api/currencies_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class CurrenciesController < BaseController
+  end
+end
