+++ app/controllers/api/flavors_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class FlavorsController < BaseController
+  end
+end
