+++ app/controllers/api/measures_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class MeasuresController < BaseController
+  end
+end
