+++ app/controllers/api/pictures_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class PicturesController < BaseController
+  end
+end
