+++ app/controllers/api/provision_dialogs_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class ProvisionDialogsController < BaseController
+  end
+end
