+++ app/controllers/api/requests_controller.rb
@@ -1,0 +1,14 @@
+module Api
+  class RequestsController < BaseController
+    def find_requests(id)
+      klass = collection_class(:requests)
+      return klass.find(id) if @auth_user_obj.admin?
+      klass.find_by!(:requester => @auth_user_obj, :id => id)
+    end
+
+    def requests_search_conditions
+      return {} if @auth_user_obj.admin?
+      {:requester => @auth_user_obj}
+    end
+  end
+end
