+++ app/controllers/api/security_groups_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class SecurityGroupsController < BaseController
+  end
+end
