+++ app/controllers/api/servers_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class ServersController < BaseController
+  end
+end
