+++ app/controllers/api/virtual_templates_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class VirtualTemplatesController < BaseController
+  end
+end
