+++ app/controllers/api/zones_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class ZonesController < BaseController
+  end
+end
